Adding strategy pattern as well as alternatives scenarios
